refactor: decouple Command components
