+++ components/Command/CommandItem.tsx
@@ -1,0 +1,32 @@
+// Copyright (c) Gridiron Survivor.
+// Licensed under the MIT License.
+
+import { forwardRef, ElementRef, ComponentPropsWithoutRef, JSX } from 'react';
+import { Command as CommandPrimitive } from 'cmdk';
+import { cn } from '@/lib/utils';
+
+/**
+ * Component that renders individual command items. 
+ * @param {ComponentPropsWithoutRef<typeof CommandPrimitive.Item>} props - Props passed to CommandPrimitive.Item
+ * @param {string} [props.className] - Additional class names for styling.
+ * @returns {JSX.Element} Rendered command item. 
+ */
+const CommandItem = forwardRef<
+  ElementRef<typeof CommandPrimitive.Item>,
+  ComponentPropsWithoutRef<typeof CommandPrimitive.Item>
+>(({ className, ...props }, ref) => (
+  <CommandPrimitive.Item
+    ref={ref}
+    className={cn(
+      'relative flex cursor-default gap-2 select-none items-center rounded-sm px-2 py-1.5 text-sm outline-none data-[disabled=true]:pointer-events-none data-[selected=true]:bg-accent data-[selected=true]:text-accent-foreground data-[disabled=true]:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0',
+      className,
+    )}
+    {...props}
+  />
+));
+
+CommandItem.displayName = CommandPrimitive.Item.displayName;
+
+export {
+  CommandItem
+};
